Properly fix float NaN this time!
There was another issue, unrelated to PDFs and perpendicular
hemisphere sampling. When `wo` was parallel with `n`, one of the cross
products in `orthonormal_basis_transform` would result in a degenerate
zero-vector. Zero vectors are bad and ruin things. Fix by not
computing an orthonormal base for (wo, n), but instead an arbitrary
base for n (using the binormal+bitangent method of both Optix's `Onb`
and PBR Book 2.2.4). As long as we use the same base for transforming
vectors into and out of the base-vector coordinate system, it doesn't
matter that it's arbitrary! Also, we seem to have gained a few FPS
doing this :)