Impl Karras's 2012 LBVH (buggy) & Give up on dummy_bvh
Why even bother with a dummy_bvh when we can just do a proper one? And
we did.

The implementation is closely based on Karras's radix-tree based LBVH
from "Maximizing Parallelism in the Construction of BVHs, Octrees,
andk-d Trees". When things got tough, we looked to both to Karras's
blog post "Thinking Parallel, Part III: Tree Construction on the GPU",
but even more we looked at Athas's futhark impl of the same alg in
github.com/athas/raytracingthenextweekinfuthark. Especially for the part
where parent "pointers" were added to the internal nodes and the bvh-ray
intersection function, we took a lot of *inspiration* from Athas's
code.

The LBVH almost works. Everything compiles and all, but some triangles
of the scene are not drawn. Who knows why.