Fix incorrectly computing AABB of triangle
The formula we used seemed correct intuitively, but apparently it was
wrong. Now we simple create an AABB for each point of the triangle,
and then return the union / the AABB that contains all three
sub-AABBs.

Also, fix float rounding error in hit_aabb. Do it more like how PBR
book does it.