Create accelerated scene once on program init
Instead of re-creating it every frame. Short build-time is indeed a
feature of the LBVH, but since we don't have any dynamics in the
scene, there's really no point in recomputing the BVH and stuff more
than once.

The FPS impact for small scenes is not big at all. Might be different
for larger scenes.