Fix caustics / MIS for BSDFs with delta distribution
The MIS in estimate_direct didn't take into account whether the PDF of
the BSDF-sample was a dirac delta function. If it is, we must assume a
light PDF of 0, as it's essentially impossible to sample that
direction from the light by chanse. Before, the light PDF was always
$1 / area$, which made the balance heuristic return a too small value
for specular reflections / refractions, so caustics became way too dark.