Rudimentary spectral ray tracing
Material color and emission are now defined with spectrums (of
descrete points, linearly interpolated). In `sample`, one of the
wavelengths 700nm (red), 500nm (green), and 400nm (blue) are chosen at
random, and in material.fut, the material is instantiated for that
wavelength. I.e., we compute lookup the wavelength in the material
spectrum and get the albedo/emission.

Radiance and BSDF-values are now mostly represented as a single f32 --
the intensity/throughput at the specific wavelength.

Next step will be to not just sample red, green, and blue, but to
sample from a distribution of arbitrary wavelengths. Maybe define the
sensor as a list of normal distributions around a wavelength, and
sample wavelength from those.