Add directional area light `frustumlight`
Seems to work pretty well. The sampling is probably not as
smart as it could be, so it converges pretty slowly. Also, note that
decreasing the "spread"-angle does not make the light brighter, rather
the opposite, since fewer points on the light source will be within
the necessary angle to light a point on a surface.