Wait to sort pixel samples from path_trace until visualize of distance
No point in sorting them when doing visual/camera path tracing, as
we're only going to average the results anyways, and that's
commutative.

Also, use bubble sort instead of radix sort. More efficient for small
lists like this, and the radix sort didn't play well with the dynamic
nature of the size of the array returned from filter.